requirements.txt
pyannote.audio==2.1.1
torch>=2.0.0
torchaudio>=2.0.0
python-dotenv>=1.0.0
 